refactor(entity): 重构行程相关实体类与服务接口
- 删除 ItineraryItems 实体及相关服务实现
- 修改 TripDays 的 dayOrder 类型为 Double
- 更新 Trips 实体字段，将 name 改为 title，preferences 和 isTemplate 字段被移除
- 新增 TripDayItems 实体以替代原 ItineraryItems 表结构
- 调整 WishlistItems 实体结构，增加 tripId、entityId 和 isPoi 字段
- 为 TripStatus 枚举添加 EnumValue 注解支持 MyBatis 映射
- 更新 Mapper 和 Service 接口定义以匹配新实体结构
- 扩展 TripDaysService 和 TripsService 接口，新增多个业务方法
- 在 WebSearchServiceImpl 中清理过时的 TODO 注释
- 更新接口文档中关于服务方法和参数的描述信息

+++ src/main/java/com/bread/traveler/entity/TripDays.java
@@ -41,7 +41,7 @@ public class TripDays implements Serializable {
      * 
      */
     @TableField(value = "day_order")
-    private Integer dayOrder;
+    private Double dayOrder;
 
     /**
      *
